refactor : workspace에 있는 useState들 template로 이동
